# openapi.yaml
openapi: 3.0.0
info:
  title: API de Registro de Asteroides
  version: 1.0.0
  description: API para registrar y monitorizar asteroides potencialmente peligrosos.
servers:
  - url: http://localhost:8080/api/v1
paths:
  /asteroides:
    post:
      summary: Crear asteroide
      security:
        - bearerAuth: []
      operationId: createAsteroid
      requestBody:
        description: Información del asteroide
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Asteroid'
      responses:
        '201':
          description: Asteroide creado con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asteroid'
        '400':
          description: Solicitud incorrecta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: Obtener todos los asteroides
      operationId: getAsteroids
      responses:
        '200':
          description: Lista de asteroides
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Asteroid'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /asteroides/{id}:
    get:
      summary: Obtener asteroide por ID
      operationId: getAsteroidById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Información del asteroide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asteroid'
        '404':
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Actualizar asteroide
      security:
        - bearerAuth: []
      operationId: updateAsteroid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Información actualizada del asteroide
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Asteroid'
      responses:
        '200':
          description: Asteroide actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asteroid'
        '400':
          description: Solicitud incorrecta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Eliminar asteroide
      security:
        - bearerAuth: []
      operationId: deleteAsteroid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Asteroide eliminado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Asteroide eliminado exitosamente
        '404':
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /register:
    post:
      summary: Registrar usuario
      operationId: registerUser
      requestBody:
        description: Información del usuario
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Usuario registrado
        '400':
          description: Solicitud incorrecta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /login:
    post:
      summary: Iniciar sesión
      operationId: loginUser
      requestBody:
        description: Credenciales del usuario
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Sesión iniciada
        '400':
          description: Solicitud incorrecta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Asteroid:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        diameter:
          type: number
        discovery_date:
          type: string
          format: date
        observations:
          type: string
        distances:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              distance:
                type: number
    Error:
      type: object
      properties:
        message:
          type: string